Improved _Beans_Anonymous_Filters.
1. Added default priority value.
2. Added default args value.
3. Added _construct() unit test.
4. Removed the _beans_add_anonymous_Filters() unit test (this is an integration test).
